+++ scripts/macros/monsterFeatures/mudMephit/mudMephit.js
@@ -1,0 +1,4 @@
+import {deathBurst} from './deathBurst.js';
+export let mudMephit = {
+    'deathBurst': deathBurst
+}