+++ scripts/macros/monsterFeatures/nightwalker/nightwalker.js
@@ -1,0 +1,10 @@
+import {annihilatingAura} from './annihilatingAura.js';
+import {enervatingFocus} from './enervatingFocus.js';
+import {fingerOfDoom} from './fingerOfDoom.js';
+import {lifeEater} from './lifeEater.js';
+export let nightwalker = {
+    'enervatingFocus': enervatingFocus,
+    'fingerOfDoom': fingerOfDoom,
+    'annihilatingAura': annihilatingAura,
+    'lifeEater': lifeEater
+}