+++ scripts/utility/effect.js
@@ -19,9 +19,10 @@ export async function itemDC(effect, updates, options, user) {
 }
 export async function fixOrigin(token, options, user) {
     if ((game.user.id !== user) || token.actorLink) return;
-    let updates = await token.actor.effects.reduce(async (updates, effect) => {
+    let updates = token.actor.effects.reduce((updates, effect) => {
         if (!effect.origin) return updates;
-        let origin = await fromUuid(effect.origin);
+        let origin = fromUuidSync(effect.origin);
+        if (!origin) return updates;
         if (!origin instanceof Item) return updates;
         let item = token.actor.items.get(origin.id);
         if (item) updates.push({'_id': effect.id, 'origin': item.uuid});